nexdom_bkup_sync/requirements.txt
aiohttp==3.10.5
uvicorn==0.30.6
fastapi==0.115.5
pyyaml==6.0.2
tenacity==9.0.0
supabase==2.5.5
boto3==1.34.40
b2sdk==2.3.0
python-dateutil==2.9.0.post0